requirements.txt
-e .
flask[async]>=3.1.0,<4.0.0
flask>=3.1.0,<4.0.0
gunicorn>=23.0.0,<24.0.0
requests==2.32.5
pypdf2==3.0.1
google-cloud>=0.34.0,<0.35.0
google-cloud-aiplatform
google-cloud-storage>=3.0.0,<4.0.0
google-cloud-pubsub>=2.28.0,<3.0.0
google-cloud-bigquery>=3.29.0,<4.0.0
google-cloud-tasks
google-genai>=1.2.0,<2.0.0
protobuf>=5.29.3,<6.0.0
paperscraper>=0.3.0,<0.4.0
nltk==3.9.1
PyMuPDF==1.25.3
tiktoken==0.9.0
bertopic==0.17.0
sci-hub==1.0.4
pandas==2.3.3
contextgem==0.19.2
openpyxl>=3.0.0
mistralai
# sentence-transformers==4.1.0
sentence-transformers
# perdu
# ilcdlib==5.1.0